Added some responsive breakpoints
Addition of some responsive breakpoints to the main css file. Still a
work in progress but there to make some corrections/adjustments to
presentation of grids on mobile views. Needs quite a lot more work as
Packery works best with item grids given fixed pixels for it's reflow -
once the screen hits the limit of the pixel size, it starts to get cut
oof.