Included stop-all-animations JS
Gave some extra API functionality to Youtube embed (so it can be called to stop by external button). Included external JS to stop all animated activity. Updated HTML to include a button which will stop all animation when clicked on.

Unfortunately have not figured out a way to stop the animation of the progress bar (and show a static image instead). Also the JS timer just disappears. Probably not the exact desired effect but this functionality (or removal of functionality) requires further discussion.